Address PR feedback
