+++ awswrangler/distributed/ray/modin/s3/_read_parquet.py
@@ -1,46 +1,15 @@
 """Modin on Ray S3 read parquet module (PRIVATE)."""
-import logging
 from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Union
 
 import modin.pandas as pd
 import pyarrow as pa
-import pyarrow.parquet
 from ray.data import read_datasource
 
-from awswrangler.distributed.ray import ray_remote
 from awswrangler.distributed.ray.datasources import ArrowParquetDatasource
 from awswrangler.distributed.ray.modin._utils import _to_modin
-from awswrangler.s3._read_parquet import _pyarrow_parquet_file_wrapper
 
 if TYPE_CHECKING:
     from mypy_boto3_s3 import S3Client
-
-
-_logger: logging.Logger = logging.getLogger(__name__)
-
-
-@ray_remote()
-def _read_parquet_metadata_file_distributed(
-    s3_client: Optional["S3Client"],
-    path: str,
-    s3_additional_kwargs: Optional[Dict[str, str]],
-    use_threads: Union[bool, int],
-    version_id: Optional[str] = None,
-    coerce_int96_timestamp_unit: Optional[str] = None,
-) -> Optional[pa.schema]:
-    fs = pyarrow.fs.S3FileSystem()
-    path_without_s3_prefix = path[len("s3://") :]
-
-    with fs.open_input_file(path_without_s3_prefix) as f:
-        pq_file = _pyarrow_parquet_file_wrapper(
-            source=f,
-            coerce_int96_timestamp_unit=coerce_int96_timestamp_unit,
-        )
-
-        if pq_file:
-            return pq_file.schema.to_arrow_schema()
-
-    return None
 
 
 def _read_parquet_distributed(  # pylint: disable=unused-argument
